Add rsmi_compute_process_gpus_get()
Given a process ID, give the device indices that process is
currently using.

Also:
* made corrections to how RSMI, amdgpu (ie, "card#") and
  KFD indicies translate from one another
* add a few missing error codes to rsmi_status_string()
* fix some formatting

+++ src/rocm_smi_counters.cc
@@ -368,7 +368,7 @@ amd::smi::evt::Event::stopCounter(void) {
   return 0;
 }
 
-static long
+static ssize_t
 readn(int fd, void *buf, size_t n) {
   ssize_t left = n;
   ssize_t bytes;
